Use block sub-package
