+++ servers/src/main/java/tachyon/master/next/FileSystemMaster.java
@@ -31,6 +31,8 @@ import tachyon.PrefixList;
 import tachyon.TachyonURI;
 import tachyon.conf.TachyonConf;
 import tachyon.master.Dependency;
+import tachyon.master.next.block.BlockId;
+import tachyon.master.next.block.BlockWorkerInfo;
 import tachyon.thrift.BlockInfoException;
 import tachyon.thrift.ClientBlockInfo;
 import tachyon.thrift.ClientDependencyInfo;
