+++ servers/src/main/java/tachyon/master/next/block/BlockId.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.master.next;
+package tachyon.master.next.block;
 
 public class BlockId {
