+++ servers/src/main/java/tachyon/master/next/block/BlockLocation.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.master.next;
+package tachyon.master.next.block;
 
 import tachyon.thrift.NetAddress;
