+++ servers/src/main/java/tachyon/master/next/block/ContainerIdGenerator.java
@@ -13,22 +13,8 @@
  * the License.
  */
 
-package tachyon.master.next;
+package tachyon.master.next.block;
 
-import java.util.concurrent.atomic.AtomicLong;
-
-public class BlockIdGenerator {
-
-  private final AtomicLong mNextContainerId;
-
-  // TODO: when needed, add functionality to create new full block ids.
-
-  public BlockIdGenerator() {
-    mNextContainerId = new AtomicLong(0);
-  }
-
-  public long getNewBlockContainerId() {
-    return mNextContainerId.getAndIncrement();
-  }
-
+public interface ContainerIdGenerator {
+  long getNewContainerId();
 }
